fopen #60

+++ src/kernel/system_call.c
@@ -14,9 +14,9 @@
 #define STDERR	0x00000002
 
 int system_call_exit(int return_value);
-int system_call_open(char const *file_name, unsigned int flags);
+unsigned int system_call_open(char const *file_name, unsigned int flags);
 #define SYSTEM_CALL_OPEN_FLAG_READ_ONLY 0x01
-int system_call_write(int file_descriptor, void const *buffer, size_t count);
+int system_call_write(unsigned int file_descriptor, void const *buffer, size_t count);
 
 int system_call(int eax, int ebx, int ecx, int edx, int esi, int edi, int ebp)
 {
@@ -31,9 +31,9 @@ int system_call(int eax, int ebx, int ecx, int edx, int esi, int edi, int ebp)
 	case SYSTEM_CALL_EXIT:
 		return system_call_exit(ebx);
 	case SYSTEM_CALL_WRITE:
-		return system_call_write(ebx, (void const *)(ecx + (unsigned int)((CommandTaskAdditional *)get_current_task()->additionals)->application_memory), (size_t)edx);
+		return system_call_write((unsigned int)ebx, (void const *)(ecx + (unsigned int)((CommandTaskAdditional *)get_current_task()->additionals)->application_memory), (size_t)edx);
 	case SYSTEM_CALL_OPEN:
-		return system_call_open((char const *)(ebx + (unsigned int)((CommandTaskAdditional *)get_current_task()->additionals)->application_memory), (unsigned int)ecx);
+		return (int)system_call_open((char const *)(ebx + (unsigned int)((CommandTaskAdditional *)get_current_task()->additionals)->application_memory), (unsigned int)ecx);
 	default:
 		ERROR(); // Invalid eax
 		return -1;
@@ -45,13 +45,13 @@ int system_call_exit(int return_value)
 	return exit_application(return_value, get_current_task()->task_status_segment.esp0);
 }
 
-int system_call_open(char const *file_name, unsigned int flags)
+unsigned int system_call_open(char const *file_name, unsigned int flags)
 {
 	if(!strcmp(file_name, "") && flags == SYSTEM_CALL_OPEN_FLAG_READ_ONLY)system_call_write(STDOUT, "Open root directory!\n", 21);
 	return 0;
 }
 
-int system_call_write(int file_descriptor, void const *buffer, size_t count)
+int system_call_write(unsigned int file_descriptor, void const *buffer, size_t count)
 {
 	Shell *shell = get_current_shell();
 	unsigned int counter = 0;
