+++ src/libs/system_call.c
@@ -10,12 +10,12 @@ int system_call_exit(int return_value)
 	return system_call(SYSTEM_CALL_EXIT, return_value, 0, 0, 0, 0, 0);
 }
 
-int system_call_open(char const *file_name, unsigned int flags)
+unsigned int system_call_open(char const *file_name, unsigned int flags)
 {
-	return system_call(SYSTEM_CALL_OPEN, (int)file_name, (int)flags, 0, 0, 0, 0);
+	return (unsigned int)system_call(SYSTEM_CALL_OPEN, (int)file_name, (int)flags, 0, 0, 0, 0);
 }
 
-int system_call_write(int file_descriptor, void const *buffer, size_t count)
+int system_call_write(unsigned int file_descriptor, void const *buffer, size_t count)
 {
 	return system_call(SYSTEM_CALL_WRITE, file_descriptor, (int)buffer, (int)count, 0, 0, 0);
 }
